+++ src/Service.php
@@ -21,9 +21,11 @@ use Embed\Http\Crawler;
 use Embed\Http\CurlClient;
 use Embed\Http\Url;
 use spicyweb\embeddedassets\adapters\akamai\Extractor as AkamaiExtractor;
+use spicyweb\embeddedassets\adapters\bluesky\Extractor as BlueskyExtractor;
 use spicyweb\embeddedassets\adapters\default\Extractor as DefaultExtractor;
 use spicyweb\embeddedassets\adapters\default\detectors\Type as TypeDetector;
 use spicyweb\embeddedassets\adapters\googlemaps\Extractor as GoogleMapsExtractor;
+use spicyweb\embeddedassets\adapters\openstreetmap\Extractor as OpenStreetMapExtractor;
 use spicyweb\embeddedassets\adapters\pbs\Extractor as PbsExtractor;
 use spicyweb\embeddedassets\adapters\sharepoint\Extractor as SharepointExtractor;
 use spicyweb\embeddedassets\errors\NotWhitelistedException;
@@ -115,8 +117,11 @@ class Service extends Component
         $clientSettings = [];
         $adapters = [
             'akamaized.net' => AkamaiExtractor::class,
+            'bsky.app' => BlueskyExtractor::class,
+            'bsky.social' => BlueskyExtractor::class,
             'pbs.org' => PbsExtractor::class,
             'nhpbs.org' => PbsExtractor::class,
+            'openstreetmap.org' => OpenStreetMapExtractor::class,
             'sharepoint.com' => SharepointExtractor::class,
         ];