+++ src/models/Settings.php
@@ -54,7 +54,11 @@ class Settings extends Model
         
         // Baidu
         'baidu.com',
-        
+
+        // Bluesky
+        'bsky.app',
+        'bsky.social',
+
         // CollegeHumor
         'collegehumor.com',
         
@@ -103,6 +107,9 @@ class Settings extends Model
         
         // Netflix
         'netflix.com',
+
+        // OpenStreetMap
+        'openstreetmap.org',
 
         // PBS and affiliates
         // TODO: add more affiliates as I find them